rename: request, response 패키지 생성 후 클래스 이동
