+++ src/main/java/webserver/DispatcherServlet.java
@@ -4,8 +4,8 @@ import controller.Controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webserver.http.HttpMethod;
-import webserver.http.HttpRequest;
-import webserver.http.HttpResponse;
+import webserver.http.request.HttpRequest;
+import webserver.http.response.HttpResponse;
 import webserver.http.HttpStatus;
 
 import java.io.OutputStream;
