+++ src/main/java/webserver/HandlerMapper.java
@@ -3,7 +3,7 @@ package webserver;
 import controller.Controller;
 import controller.annotaion.GetMapping;
 import controller.annotaion.PostMapping;
-import webserver.http.HttpRequest;
+import webserver.http.request.HttpRequest;
 
 import java.lang.reflect.Method;
 import java.util.HashMap;
