+++ src/main/java/webserver/http/request/HttpRequest.java
@@ -1,7 +1,9 @@
-package webserver.http;
+package webserver.http.request;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.http.HttpMethod;
+import webserver.http.HttpMime;
 
 import java.io.BufferedReader;
 import java.io.IOException;
