+++ src/main/java/webserver/http/response/HttpResponse.java
@@ -1,7 +1,9 @@
-package webserver.http;
+package webserver.http.response;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.http.HttpMime;
+import webserver.http.HttpStatus;
 
 import java.io.DataOutputStream;
 import java.io.File;
